requirements.txt
Django>=5.0,<6.0
django-htmx>=1.17.0
anthropic>=0.18.0
openai>=1.12.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# PDF Generation
reportlab>=4.0.0

# Web Research for Profile Enrichment
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0

# Claude Agent SDK (uses Claude Code Max subscription)
claude-agent-sdk>=0.1.0

# Logging
python-json-logger>=2.0.0

# Contact Verification
dnspython>=2.4.0
phonenumbers>=8.13.0

# Production
gunicorn>=21.0.0
whitenoise>=6.6.0
